requirements.txt
pexpect
textual
ruamel.yamlemail-validator>=2.0.0
